+++ pidev-backend/src/main/java/tn/esprit/pidev/services/ressources/IGestionRessource.java
@@ -1,13 +1,13 @@
 package tn.esprit.pidev.services.ressources;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import tn.esprit.pidev.entities.user.User;
-import tn.esprit.pidev.entities.ressources.Categorie;
-import tn.esprit.pidev.entities.ressources.Options;
-import tn.esprit.pidev.entities.ressources.Ressource;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.web.multipart.MultipartFile;
+import tn.esprit.pidev.entities.ressources.Categorie;
+import tn.esprit.pidev.entities.ressources.Options;
+import tn.esprit.pidev.entities.ressources.Ressource;
+import tn.esprit.pidev.entities.user.User;
 
 import java.io.IOException;
 import java.util.List;
@@ -18,7 +18,7 @@ public interface IGestionRessource {
     List<Ressource> getRessourceByCategory(Categorie categorie, Long topicId);
 
     @Transactional
-    Ressource addRessource(MultipartFile file, Ressource ressource, Long userId, String topicName) throws IOException;
+    Ressource addRessource(MultipartFile file, Ressource ressource, Integer userId, String topicName) throws IOException;
 
     @Transactional
     Ressource updateRessource(Long ressourceId, Ressource updatedRessource, MultipartFile newFile) throws IOException;
@@ -56,5 +56,5 @@ public interface IGestionRessource {
 
     int countResourcesByTopicId(Long topicId);
 
-    List<Ressource> getResourcesByUserId(Long userId);
+    List<Ressource> getResourcesByUserId(Integer userId);
 }
